refactor: move coupling_cap/assert_grid_covers_coupling to specs
Both are pure functions over a TiltSegmentUnion policy + a scalar grid radius,
with no preprocess-specific dependencies. Homing them in specs (next to the
value-type) lets config and preprocess both depend downward on specs, removing
the config -> preprocess layering inversion and the lazy import the config-load
guard needed to break that cycle.

Pure code motion, no behavior change: preprocess.coupling now imports
coupling_cap from specs (used by tilt_segment_coupling); config.schema and
fit_orientation import assert_grid_covers_coupling from specs; tests updated.

Tests: full tests/unit green (546 passed, 3 skipped)
diffBloch_private analog: none (public-only layering cleanup atop 955fa234 port)
